fix: Migrate from deprecated Detekt API
In the change to detekt v1.24.3, several detekt APIs were deprecated,
but this was not done correctly, and some APIs were actually removed!

Further, if the detekt config file includes formatting settings,
detekt-cli requires the detekt formatting plugin to be on the classpath
(another breaking change). Formatting is done with ktfmt so we just
remove the formatting section from the detekt config.